moved walltalker into service tabs
