+++ src/game/components/GameSettings.tsx
@@ -13,7 +13,6 @@ import {
   PaceSettings,
   PlayerSettings,
   VibratorSettings,
-  WalltalkerSettings,
 } from '../../settings';
 import { GamePhase, useGameValue, useSendMessage } from '../GameProvider';
 import { useFullscreen, useLooping } from '../../utils';
@@ -60,7 +59,6 @@ const GameSettingsDialog: React.FC<GameSettingsDialogProps> = props => {
           <ClimaxSettings />
           <BoardSettings />
           <VibratorSettings />
-          <WalltalkerSettings />
           <ImageSettings />
         </StyledGameSettingsDialog>
       }
